requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
future==0.16.0
idna==2.6
markovify==0.6.0
nltk==3.2.4
oauthlib==2.0.2
python-telegram-bot==7.0.1
requests==2.18.4
requests-oauthlib==0.8.0
six==1.10.0
tweepy==3.5.0
Unidecode==0.4.21
urllib3==1.22
